+++ moban/engine_factory.py
@@ -4,9 +4,23 @@ from collections import defaultdict
 import moban.utils as utils
 import moban.constants as constants
 import moban.exceptions as exceptions
-from moban.extensions import LibraryManager
+from lml.plugin import PluginManager
 
-LIBRARIES = LibraryManager()
+
+class EngineFactory(PluginManager):
+    def __init__(self):
+        super(EngineFactory, self).__init__(
+            constants.TEMPLATE_ENGINE_EXTENSION
+        )
+
+    def get_engine(self, template_type):
+        return self.load_me_now(template_type)
+
+    def all_types(self):
+        return list(self.registry.keys())
+
+    def raise_exception(self, key):
+        raise exceptions.NoThirdPartyEngine(key)
 
 
 class Context(object):
@@ -68,44 +82,6 @@ def _append_to_array_item_to_dictionary_key(adict, key, array_item):
         )
     else:
         adict[key].append(array_item)
-
-
-def expand_template_directories(dirs):
-    if not isinstance(dirs, list):
-        dirs = [dirs]
-
-    for directory in dirs:
-        yield expand_template_directory(directory)
-
-
-def expand_template_directory(directory):
-    translated_directory = None
-    if ":" in directory:
-        library_or_repo_name, relative_path = directory.split(":")
-        potential_repo_path = os.path.join(
-            utils.get_moban_home(), library_or_repo_name
-        )
-        if os.path.exists(potential_repo_path):
-            # expand repo template path
-            if relative_path:
-                translated_directory = os.path.join(
-                    potential_repo_path, relative_path
-                )
-            else:
-                translated_directory = potential_repo_path
-        else:
-            # expand pypi template path
-            library_path = LIBRARIES.resource_path_of(library_or_repo_name)
-            if relative_path:
-                translated_directory = os.path.join(
-                    library_path, relative_path
-                )
-            else:
-                translated_directory = library_path
-    else:
-        # local template path
-        translated_directory = os.path.abspath(directory)
-    return translated_directory
 
 
 def verify_the_existence_of_directories(dirs):
