+++ tests/test_engine.py
@@ -7,8 +7,9 @@ from mock import patch
 from nose.tools import eq_, raises
 from moban.engine import Engine
 from moban.extensions import jinja_global
-from moban.engine_factory import Context, expand_template_directories
-from moban.engine import ENGINES
+from moban.engine_factory import Context
+from moban.plugins import expand_template_directories
+from moban.plugins import ENGINES
 from moban.engine_handlebars import EngineHandlebars
